Mirror tty when in tty and not no-color passed
 - Fixes https://github.com/buildpacks/pack/issues/493
