+++ internal/image/fetcher.go
@@ -11,15 +11,13 @@ import (
 	"github.com/buildpacks/imgutil/local"
 	"github.com/buildpacks/imgutil/remote"
 	"github.com/buildpacks/lifecycle/auth"
+	"github.com/buildpacks/pack/internal/style"
+	"github.com/buildpacks/pack/logging"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/client"
 	"github.com/docker/docker/pkg/jsonmessage"
 	"github.com/google/go-containerregistry/pkg/authn"
 	"github.com/pkg/errors"
-	"golang.org/x/crypto/ssh/terminal"
-
-	"github.com/buildpacks/pack/internal/style"
-	"github.com/buildpacks/pack/logging"
 )
 
 type Fetcher struct {
@@ -90,7 +88,7 @@ func (f *Fetcher) pullImage(ctx context.Context, imageID string) error {
 	}
 
 	writer := logging.GetWriterForLevel(f.logger, logging.InfoLevel)
-	termFd, isTerm := isTerminal(writer)
+	termFd, isTerm := logging.IsTerminal(writer)
 
 	err = jsonmessage.DisplayJSONMessagesStream(rc, &colorizedWriter{writer}, termFd, isTerm, nil)
 	if err != nil {
@@ -98,20 +96,6 @@ func (f *Fetcher) pullImage(ctx context.Context, imageID string) error {
 	}
 
 	return rc.Close()
-}
-
-func isTerminal(w io.Writer) (uintptr, bool) {
-	type descriptor interface {
-		Fd() uintptr
-	}
-
-	if f, ok := w.(descriptor); ok {
-		termFd := f.Fd()
-		isTerm := terminal.IsTerminal(int(termFd))
-		return termFd, isTerm
-	}
-
-	return 0, false
 }
 
 func registryAuth(ref string) (string, error) {
